Fake_news_authenticator__/Fake-News-Detector/requirements.txt
# Core Dependencies
streamlit>=1.28.0
streamlit-option-menu>=0.3.2

# Data Processing & ML
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Web & API
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0

# Environment & Configuration
python-dotenv>=1.0.0
